requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=0.1=cpu_0
blas=1.0=mkl
ca-certificates=2021.5.25=haa95532_1
certifi=2021.5.30=py38haa95532_0
cffi=1.14.5=py38hcd4344a_0
icc_rt=2019.0.0=h0cc432a_1
intel-openmp=2019.4=245
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
liblapack=3.9.0=5_hd5c7e75_netlib
libmklml=2019.0.5=haa95532_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
mkl=2019.4=245
mkl-service=2.3.0=py38h196d8e1_0
msys2-conda-epoch=20160418=1
ninja=1.10.2=h6d14046_1
numpy=1.20.3=py38h09042cb_0
openssl=1.1.1k=h2bbff1b_0
pandas=1.2.4=py38hd77b12b_0
patsy=0.5.1=py38_0
pip=21.1.1=py38haa95532_0
pycparser=2.20=py_2
python=3.8.10=hdbf39b2_7
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.6.0=cpu_py38h538a6d7_0
pytz=2021.1=pyhd3eb1b0_0
scipy=1.6.2=py38h14eb087_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=py38haa95532_0
sqlite=3.35.4=h2bbff1b_0
statsmodels=0.12.2=py38h2bbff1b_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
